+++ plugins/woocommerce-admin/packages/date/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Take into account leap year in calculating `getLastPeriod`.
+
 # 2.1.0
 
 -   Update to @wordpress/eslint coding standards.